% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cr_channels}
\alias{cr_channels}
\title{List channels in a \code{cr_experiment}}
\usage{
cr_channels(experiment)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}
}
\value{
A character vector of channel names.
}
\description{
List channels in a \code{cr_experiment}
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 10)
cr_channels(exp)
}
