% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_figure_size}
\alias{cr_figure_size}
\title{Standard deterministic figure dimensions}
\usage{
cr_figure_size(
  size = c("single", "double", "report", "square"),
  units = c("mm", "in")
)
}
\arguments{
\item{size}{One of \code{single}, \code{double}, \code{report}, or \code{square}.}

\item{units}{\code{"mm"} or \code{"in"}.}
}
\description{
Standard deterministic figure dimensions
}
