% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_format_percent}
\alias{cr_format_percent}
\title{Format proportions as percentages}
\usage{
cr_format_percent(x, digits = 1)
}
\arguments{
\item{x}{Numeric proportions.}

\item{digits}{Decimal places.}
}
\description{
Format proportions as percentages
}
