% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_format_pvalue}
\alias{cr_format_pvalue}
\title{Format p-values without displaying calculated zero}
\usage{
cr_format_pvalue(x, digits = 3)
}
\arguments{
\item{x}{Numeric p-values.}

\item{digits}{Decimal places determining the reporting threshold.}
}
\description{
Format p-values without displaying calculated zero
}
