% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_dose_response}
\alias{cr_plot_dose_response}
\title{Dose-response plot}
\usage{
cr_plot_dose_response(fit)
}
\arguments{
\item{fit}{A \code{cr_dose_response}.}
}
\value{
A ggplot2 object.
}
\description{
Dose-response plot
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
exp$design$dose <- ifelse(exp$design$treatment == "CompoundA_high",
                          500, exp$design$dose)
fit <- cr_dose_response(exp, channel = "marker_1", model = "4pl")
cr_plot_dose_response(fit)
}
