% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_histogram}
\alias{cr_plot_histogram}
\title{Histogram of channel intensity}
\usage{
cr_plot_histogram(
  experiment,
  channel,
  group_by = "treatment",
  facet_by = NULL,
  log_x = TRUE
)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel}{Channel name.}

\item{group_by}{Design column used for colour.}

\item{facet_by}{Design column used for facetting (or \code{NULL}).}

\item{log_x}{Log-transform the x-axis.}
}
\value{
A \code{ggplot} object.
}
\description{
Histogram of channel intensity
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_plot_histogram(exp, "marker_1")
}
\seealso{
Other visualisation:
\code{\link[=cr_plot_comparison]{cr_plot_comparison()}},
\code{\link[=cr_plot_dose_response]{cr_plot_dose_response()}},
\code{\link[=cr_plot_effect_sizes]{cr_plot_effect_sizes()}},
\code{\link[=cr_plot_foldchange]{cr_plot_foldchange()}},
\code{\link[=cr_plot_heatmap]{cr_plot_heatmap()}},
\code{\link[=cr_plot_intensity]{cr_plot_intensity()}},
\code{\link[=cr_plot_plate]{cr_plot_plate()}},
\code{\link[=cr_plot_qc]{cr_plot_qc()}},
\code{\link[=cr_plot_roc]{cr_plot_roc()}},
\code{\link[=cr_plot_scatter]{cr_plot_scatter()}},
\code{\link[=cr_plot_spatial]{cr_plot_spatial()}},
\code{\link[=cr_plot_timeline]{cr_plot_timeline()}}
}
\concept{visualisation}
