% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_plot_qc_summary}
\alias{cr_plot_qc_summary}
\title{Compact text-first QC status plot}
\usage{
cr_plot_qc_summary(qc, mode = c("colour", "grayscale"))
}
\arguments{
\item{qc}{A report QC table.}

\item{mode}{Colour mode.}
}
\description{
Compact text-first QC status plot
}
