% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_roc}
\alias{cr_plot_roc}
\title{ROC curve plot}
\usage{
cr_plot_roc(result)
}
\arguments{
\item{result}{A single \code{cr_result} or a list of such results.}
}
\value{
A ggplot2 object.
}
\description{
ROC curve plot
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
res <- cr_logistic(exp, "marker_1", "CompoundA_high", "Untreated")
cr_plot_roc(res)
}
