% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_spatial}
\alias{cr_plot_spatial}
\title{Spatial scatter of cells in a well}
\usage{
cr_plot_spatial(experiment, channel, well = NULL)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel}{Channel used for colouring points.}

\item{well}{Spatial unit (well or slide ID) to plot. If \code{NULL},
the first well is chosen.}
}
\value{
A ggplot2 object.
}
\description{
Spatial scatter of cells in a well
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_plot_spatial(exp, "marker_1")
}
