% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_timeline}
\alias{cr_plot_timeline}
\title{Time-course line plot}
\usage{
cr_plot_timeline(experiment, timepoint_var, channel, group_by = "treatment")
}
\arguments{
\item{experiment}{A \code{cr_experiment}. Design table must have a
time variable column.}

\item{timepoint_var}{Name of the time variable column.}

\item{channel}{Channel to plot.}

\item{group_by}{Grouping variable (colour).}
}
\value{
A ggplot2 object.
}
\description{
Time-course line plot
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
exp$design$timepoint <- rep(c(0, 6, 12, 24), length.out = nrow(exp$design))
cr_plot_timeline(exp, "timepoint", "marker_1")
}
