% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-spec.R
\name{cr_print_report_spec}
\alias{cr_print_report_spec}
\title{Print a laboratory report specification}
\usage{
cr_print_report_spec(x, ...)
}
\arguments{
\item{x}{A \code{cr_report_spec}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a laboratory report specification
}
