% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{cr_qc_intensity}
\alias{cr_qc_intensity}
\title{Gate cells by intensity}
\usage{
cr_qc_intensity(experiment, channel, min_intensity = NA, max_intensity = NA)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel}{Name of the channel column in \code{cells}.}

\item{min_intensity, max_intensity}{Lower and upper bounds. Leave
\code{NA} to skip either.}
}
\value{
A modified \code{cr_experiment}.
}
\description{
Removes cells whose intensity on a given channel is outside the
interval \verb{[min_intensity, max_intensity]}.
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_qc_intensity(exp, channel = "DAPI", min_intensity = 50)
}
