% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{cr_read_design}
\alias{cr_read_design}
\title{Read experimental design from CSV or Excel}
\usage{
cr_read_design(path)
}
\arguments{
\item{path}{Path to file. \code{.csv}, \code{.tsv} and \code{.xlsx} are
supported.}
}
\value{
A tibble of design information.
}
\description{
Read experimental design from CSV or Excel
}
\examples{
d <- tempfile("cr_design_"); dir.create(d)
files <- cr_example_files(d)
design_file <- grep("design", files, value = TRUE)[1]
design <- cr_read_design(design_file)
head(design)
}
