% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lab.R
\name{cr_report_hash}
\alias{cr_report_hash}
\title{Hash report-relevant structured data}
\usage{
cr_report_hash(x, algorithm = "sha256")
}
\arguments{
\item{x}{Any R object.}

\item{algorithm}{Algorithm supported by \code{\link[digest:digest]{digest::digest()}}, default SHA-256.}
}
\description{
Hash report-relevant structured data
}
