% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cr_rowcol_to_well}
\alias{cr_rowcol_to_well}
\title{Convert row and column indices to well IDs}
\usage{
cr_rowcol_to_well(row, col, pad = 2)
}
\arguments{
\item{row}{Integer vector of row numbers (1 = A, 2 = B, ...).}

\item{col}{Integer vector of column numbers.}

\item{pad}{Number of digits to pad the column number to
(default 2, yielding \code{"A01"} rather than \code{"A1"}).}
}
\value{
A character vector of well IDs.
}
\description{
Convert row and column indices to well IDs
}
\examples{
cr_rowcol_to_well(c(1, 2, 8), c(1, 5, 12))
}
