% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_scale_colour}
\alias{cr_scale_colour}
\alias{cr_scale_fill}
\alias{cr_shape_scale}
\alias{cr_linetype_scale}
\title{Manual colour scale from the cellreportR visual system}
\usage{
cr_scale_colour(values = NULL, mode = c("colour", "grayscale"), ...)

cr_scale_fill(values = NULL, mode = c("colour", "grayscale"), ...)

cr_shape_scale(values = NULL, ...)

cr_linetype_scale(values = NULL, ...)
}
\arguments{
\item{values}{Optional named values.}

\item{mode}{Colour mode.}

\item{...}{Passed to \code{ggplot2::scale_colour_manual()}.}
}
\description{
Manual colour scale from the cellreportR visual system
}
