% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-publication.R
\name{cr_scale_diverging}
\alias{cr_scale_diverging}
\title{Continuous diverging scale}
\usage{
cr_scale_diverging(
  midpoint = 0,
  aesthetics = c("colour", "fill"),
  mode = c("colour", "grayscale"),
  ...
)
}
\arguments{
\item{midpoint}{Explicit neutral midpoint.}

\item{aesthetics}{\code{"colour"} or \code{"fill"}.}

\item{mode}{Colour mode.}

\item{...}{Passed to the gradient scale.}
}
\description{
Continuous diverging scale
}
