% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{cr_validate_experiment}
\alias{cr_validate_experiment}
\title{Validate a \code{cr_experiment}}
\usage{
cr_validate_experiment(x)
}
\arguments{
\item{x}{A \code{cr_experiment}.}
}
\value{
\code{TRUE} invisibly on success. On failure an informative
error is raised.
}
\description{
Performs structural checks on a \code{cr_experiment}. Called
automatically by \code{\link[=cr_build_experiment]{cr_build_experiment()}} but can also be used to
verify that manual modifications have not broken the object.
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 10)
cr_validate_experiment(exp)
}
