% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.cr_experiment}
\alias{print.cr_experiment}
\title{Print method for \code{cr_experiment}}
\usage{
\method{print}{cr_experiment}(x, ...)
}
\arguments{
\item{x}{A \code{cr_experiment}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for \code{cr_experiment}
}
