% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.cr_report}
\alias{print.cr_report}
\title{Print method for \code{cr_report}}
\usage{
\method{print}{cr_report}(x, ...)
}
\arguments{
\item{x}{A \code{cr_report}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for \code{cr_report}
}
