% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.cr_result}
\alias{print.cr_result}
\title{Print method for \code{cr_result}}
\usage{
\method{print}{cr_result}(x, ...)
}
\arguments{
\item{x}{A \code{cr_result}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for \code{cr_result}
}
