% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.cr_experiment}
\alias{summary.cr_experiment}
\title{Summary method for \code{cr_experiment}}
\usage{
\method{summary}{cr_experiment}(object, ...)
}
\arguments{
\item{object}{A \code{cr_experiment}.}

\item{...}{Unused.}
}
\value{
A tibble summarising the experiment (invisibly printed).
}
\description{
Summary method for \code{cr_experiment}
}
