% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.cr_report}
\alias{summary.cr_report}
\title{Summary method for \code{cr_report}}
\usage{
\method{summary}{cr_report}(object, ...)
}
\arguments{
\item{object}{A \code{cr_report}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summary method for \code{cr_report}
}
