% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.cr_result}
\alias{summary.cr_result}
\title{Summary method for \code{cr_result}}
\usage{
\method{summary}{cr_result}(object, ...)
}
\arguments{
\item{object}{A \code{cr_result}.}

\item{...}{Unused.}
}
\value{
The result list, invisibly.
}
\description{
Summary method for \code{cr_result}
}
