% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lab.R
\name{cr_export_report_spec}
\alias{cr_export_report_spec}
\title{Export a report specification as versioned JSON}
\usage{
cr_export_report_spec(spec, path, pretty = TRUE)
}
\arguments{
\item{spec}{A \code{cr_report_spec}.}

\item{path}{Output JSON path.}

\item{pretty}{Pretty-print JSON.}
}
\description{
Export a report specification as versioned JSON
}
