% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lab.R
\name{cr_import_report_spec}
\alias{cr_import_report_spec}
\title{Import a versioned report specification from JSON}
\usage{
cr_import_report_spec(path, validate = TRUE)
}
\arguments{
\item{path}{Input JSON path.}

\item{validate}{Validate the reconstructed specification.}
}
\description{
Import a versioned report specification from JSON
}
