% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-results.R
\name{cr_plot_dose_response}
\alias{cr_plot_dose_response}
\title{Dose-response plot}
\usage{
cr_plot_dose_response(fit)
}
\arguments{
\item{fit}{A \code{cr_dose_response}.}
}
\value{
A \code{ggplot} object.
}
\description{
Dose-response plot
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
exp$design$dose <- ifelse(exp$design$treatment == "CompoundA_high",
                          500, exp$design$dose)
fit <- cr_dose_response(exp, channel = "marker_1", model = "4pl")
cr_plot_dose_response(fit)
}
\seealso{
Other visualisation:
\code{\link[=cr_plot_comparison]{cr_plot_comparison()}},
\code{\link[=cr_plot_effect_sizes]{cr_plot_effect_sizes()}},
\code{\link[=cr_plot_foldchange]{cr_plot_foldchange()}},
\code{\link[=cr_plot_heatmap]{cr_plot_heatmap()}},
\code{\link[=cr_plot_histogram]{cr_plot_histogram()}},
\code{\link[=cr_plot_intensity]{cr_plot_intensity()}},
\code{\link[=cr_plot_plate]{cr_plot_plate()}},
\code{\link[=cr_plot_qc]{cr_plot_qc()}},
\code{\link[=cr_plot_roc]{cr_plot_roc()}},
\code{\link[=cr_plot_scatter]{cr_plot_scatter()}},
\code{\link[=cr_plot_spatial]{cr_plot_spatial()}},
\code{\link[=cr_plot_timeline]{cr_plot_timeline()}}
}
\concept{visualisation}
