% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_histogram}
\alias{cr_plot_histogram}
\title{Histogram of channel intensity}
\usage{
cr_plot_histogram(
  experiment,
  channel,
  group_by = "treatment",
  facet_by = NULL,
  log_x = TRUE
)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel}{Channel name.}

\item{group_by}{Design column used for colour.}

\item{facet_by}{Design column used for facetting (or \code{NULL}).}

\item{log_x}{Log-transform the x-axis.}
}
\value{
A ggplot2 object.
}
\description{
Histogram of channel intensity
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_plot_histogram(exp, "marker_1")
}
