% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_scatter}
\alias{cr_plot_scatter}
\title{Biaxial scatter plot of two channels}
\usage{
cr_plot_scatter(
  experiment,
  channel_x,
  channel_y,
  color_by = "treatment",
  log_x = TRUE,
  log_y = TRUE
)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel_x}{X-axis channel.}

\item{channel_y}{Y-axis channel.}

\item{color_by}{Design column used for colour. Default
\code{"treatment"}.}

\item{log_x, log_y}{Log-transform the respective axis.}
}
\value{
A \code{ggplot} object.
}
\description{
Biaxial scatter plot of two channels
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_plot_scatter(exp, "DAPI", "marker_1")
}
\seealso{
Other visualisation:
\code{\link[=cr_plot_comparison]{cr_plot_comparison()}},
\code{\link[=cr_plot_dose_response]{cr_plot_dose_response()}},
\code{\link[=cr_plot_effect_sizes]{cr_plot_effect_sizes()}},
\code{\link[=cr_plot_foldchange]{cr_plot_foldchange()}},
\code{\link[=cr_plot_heatmap]{cr_plot_heatmap()}},
\code{\link[=cr_plot_histogram]{cr_plot_histogram()}},
\code{\link[=cr_plot_intensity]{cr_plot_intensity()}},
\code{\link[=cr_plot_plate]{cr_plot_plate()}},
\code{\link[=cr_plot_qc]{cr_plot_qc()}},
\code{\link[=cr_plot_roc]{cr_plot_roc()}},
\code{\link[=cr_plot_spatial]{cr_plot_spatial()}},
\code{\link[=cr_plot_timeline]{cr_plot_timeline()}}
}
\concept{visualisation}
