% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cr_plot_spatial}
\alias{cr_plot_spatial}
\title{Spatial scatter of cells in a well}
\usage{
cr_plot_spatial(experiment, channel, well = NULL)
}
\arguments{
\item{experiment}{A \code{cr_experiment}.}

\item{channel}{Channel used for colouring points.}

\item{well}{Spatial unit (well or slide ID) to plot. If \code{NULL},
the first well is chosen.}
}
\value{
A \code{ggplot} object.
}
\description{
Spatial scatter of cells in a well
}
\examples{
exp <- cr_example_experiment(seed = 1, n_cells_per_well = 30)
cr_plot_spatial(exp, "marker_1")
}
\seealso{
Other visualisation:
\code{\link[=cr_plot_comparison]{cr_plot_comparison()}},
\code{\link[=cr_plot_dose_response]{cr_plot_dose_response()}},
\code{\link[=cr_plot_effect_sizes]{cr_plot_effect_sizes()}},
\code{\link[=cr_plot_foldchange]{cr_plot_foldchange()}},
\code{\link[=cr_plot_heatmap]{cr_plot_heatmap()}},
\code{\link[=cr_plot_histogram]{cr_plot_histogram()}},
\code{\link[=cr_plot_intensity]{cr_plot_intensity()}},
\code{\link[=cr_plot_plate]{cr_plot_plate()}},
\code{\link[=cr_plot_qc]{cr_plot_qc()}},
\code{\link[=cr_plot_roc]{cr_plot_roc()}},
\code{\link[=cr_plot_scatter]{cr_plot_scatter()}},
\code{\link[=cr_plot_timeline]{cr_plot_timeline()}}
}
\concept{visualisation}
